Move PsbtInputSatisfier to a new satisfy submodule
Move the `PsbtInputSatisfier` to a newly create `satisfy` submodule and
re-export the type from the original module.

Code move only, no logic changes.
